+++ packages/components-angular/projects/components-wrapper/src/lib/bundle.d.ts
@@ -1591,6 +1591,5 @@ declare namespace LocalJSX {
 		"p-textarea-wrapper": PTextareaWrapper;
 	}
 }
-export declare const componentsReady: () => Promise<void>;
 
 export {};