+++ packages/components-react/projects/components-wrapper/src/mocks/p-text-list-mock.tsx
@@ -1,4 +1,0 @@
-//@ts-nocheck
-import React from 'react';
-
-export const PTextList = (props) => <p-text-list {...props} />;