+++ packages/components-react/tests/jsdom-polyfill/specs/textarea.spec.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import {
+  componentsReady, PTextareaWrapper
+} from '@porsche-design-system/components-react';
+import { render } from '@testing-library/react';
+
+const Sample = (): JSX.Element => {
+  return (
+    <>
+      <PTextareaWrapper label="Some label" data-testid="host">
+        <textarea />
+      </PTextareaWrapper>
+    </>
+  );
+};
+
+describe('PTextareaWrapper', () => {
+  it('should have initialized shadow dom', async () => {
+    const { getByTestId } = render(<Sample />);
+
+    await componentsReady();
+
+    expect(getByTestId('host').shadowRoot).not.toBeNull();
+  });
+});