+++ packages/components-js/tests/e2e/specs/tabs-bar.e2e.ts
@@ -9,7 +9,7 @@ import {
   reattachElement,
   selectNode, setAttribute,
   setContentWithDesignSystem, waitForInheritedCSSTransition, expectedStyleOnFocus,
-  waitForStencilLifecycle
+  waitForStencilLifecycle,
 } from '../helpers';
 
 export const CSS_ANIMATION_DURATION = 1000;
@@ -391,7 +391,7 @@ describe('tabs-bar', () => {
       expect(eventCounter).toBe(3);
     });
 
-    it('should not dispatch event initially', async () => {
+    it('should not dispatch event initially with valid activeTabIndex', async () => {
       const COUNTER_KEY = 'pdsEventCounter';
       await setContentWithDesignSystem(page, ''); // empty page
 
@@ -483,7 +483,10 @@ describe('tabs-bar', () => {
       page.on('console', (msg) => {
         consoleMessages.push(msg);
         if (msg.type() === 'error') {
-          console.log(msg.args()[0]['_remoteObject'].description);
+          const { description } = msg.args()[0]['_remoteObject'];
+          if (description) {
+            console.log(description);
+          }
         }
       });
 
@@ -512,7 +515,10 @@ describe('tabs-bar', () => {
       page.on('console', (msg) => {
         consoleMessages.push(msg);
         if (msg.type() === 'error') {
-          console.log(msg.args()[0]['_remoteObject'].description);
+          const { description } = msg.args()[0]['_remoteObject'];
+          if (description) {
+            console.log(description);
+          }
         }
       });