+++ packages/components-react/projects/components-wrapper/src/mocks/p-content-wrapper-mock.tsx
@@ -1,4 +1,0 @@
-//@ts-nocheck
-import React from 'react';
-
-export const PContentWrapper = (props) => <p-content-wrapper {...props} />;