+++ packages/components-react/tests/jsdom-polyfill/specs/button-pure.spec.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import {
+  componentsReady, PButtonPure
+} from '@porsche-design-system/components-react';
+import { render } from '@testing-library/react';
+
+const Sample = (): JSX.Element => {
+  return (
+    <>
+      <PButtonPure data-testid="host">Some label</PButtonPure>
+    </>
+  );
+};
+
+describe('PButtonPure', () => {
+  it('should have initialized shadow dom', async () => {
+    const { getByTestId } = render(<Sample />);
+
+    await componentsReady();
+
+    expect(getByTestId('host').shadowRoot).not.toBeNull();
+  });
+});